# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
...
  /job-postings:
    get:
      operationId: listJobPostings
      tags: [recruitment]
      summary: List job postings
      responses:
        "200":
          description: List of job postings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobPosting"
    post:
      operationId: createJobPosting
      tags: [recruitment]
      summary: Create a new job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPostingInput"
      responses:
        "201":
          description: Job posting created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPosting"
  /job-postings/{id}:
    patch:
      operationId: updateJobPosting
      tags: [recruitment]
      summary: Update job posting by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPostingUpdate"
      responses:
        "200":
          description: Job posting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPosting"
  /job-postings/{id}/applications:
    get:
      operationId: listApplications
      tags: [recruitment]
      summary: List applications for a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [recruitment]
      summary: Create a new application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [recruitment]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /events:
    get:
      operationId: listEvents
      tags: [calendar]
      summary: List events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [calendar]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [calendar]
      summary: Update event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [calendar]
      summary: Delete event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteEvent200"
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
components:
  schemas:
...
    Goal:
      type: object
      required: [id, employeeId, title, category, progress, status, createdAt]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: string }
        progress: { type: integer }
        status: { type: string }
        dueDate: { type: ["string", "null"] }
        createdAt: { type: string }
    GoalInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        progress: { type: integer }
        status: { type: string }
        dueDate: { type: string }
    GoalUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        progress: { type: integer }
        status: { type: string }
        dueDate: { type: string }
...
    JobPosting:
      type: object
      required: [id, title, department, location, type, description, status, postedById, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        department: { type: string }
        location: { type: string }
        type: { type: string }
        description: { type: string }
        requirements: { type: ["string", "null"] }
        salaryMin: { type: ["integer", "null"] }
        salaryMax: { type: ["integer", "null"] }
        status: { type: string }
        postedById: { type: integer }
        createdAt: { type: string }
    JobPostingInput:
      type: object
      required: [title, department, description]
      properties:
        title: { type: string }
        department: { type: string }
        location: { type: string }
        type: { type: string }
        description: { type: string }
        requirements: { type: string }
        salaryMin: { type: integer }
        salaryMax: { type: integer }
    JobPostingUpdate:
      type: object
      properties:
        title: { type: string }
        department: { type: string }
        location: { type: string }
        type: { type: string }
        description: { type: string }
        requirements: { type: string }
        salaryMin: { type: integer }
        salaryMax: { type: integer }
        status: { type: string }
    Application:
      type: object
      required: [id, jobPostingId, applicantName, applicantEmail, status, appliedAt]
      properties:
        id: { type: integer }
        jobPostingId: { type: integer }
        applicantName: { type: string }
        applicantEmail: { type: string }
        phone: { type: ["string", "null"] }
        resumeUrl: { type: ["string", "null"] }
        coverLetter: { type: ["string", "null"] }
        status: { type: string }
        notes: { type: ["string", "null"] }
        appliedAt: { type: string }
    ApplicationInput:
      type: object
      required: [applicantName, applicantEmail]
      properties:
        applicantName: { type: string }
        applicantEmail: { type: string }
        phone: { type: string }
        coverLetter: { type: string }
        notes: { type: string }
    ApplicationUpdate:
      type: object
      properties:
        status: { type: string }
        notes: { type: string }
...
    Event:
      type: object
      required: [id, title, type, startDate, endDate, allDay, organizerId, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        type: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        location: { type: ["string", "null"] }
        allDay: { type: boolean }
        organizerId: { type: integer }
        createdAt: { type: string }
    EventInput:
      type: object
      required: [title, startDate, endDate]
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        location: { type: string }
        allDay: { type: boolean }
    EventUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        location: { type: string }
        allDay: { type: boolean }
    DeleteEvent200:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
    Asset:
      type: object
      required: [id, name, type, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        serialNumber: { type: ["string", "null"] }
        assignedToId: { type: ["integer", "null"] }
        status: { type: string }
        purchaseDate: { type: ["string", "null"] }
        purchaseValue: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    AssetInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string }
        serialNumber: { type: string }
        assignedToId: { type: integer }
        status: { type: string }
        purchaseDate: { type: string }
        purchaseValue: { type: number }
        notes: { type: string }
    AssetUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        serialNumber: { type: string }
        assignedToId: { type: integer }
        status: { type: string }
        notes: { type: string }