# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  name: string;
  type: string;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  assignedToId?: number | null;
  status: string;
  /** @nullable */
  purchaseDate?: string | null;
  /** @nullable */
  purchaseValue?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}