# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  title: string;
  description?: string;
  category?: string;
  progress?: number;
  status?: string;
  dueDate?: string;
}